put functions into TodayItem
